+++ Backend/main/resources/poemas.py
@@ -6,17 +6,25 @@ from datetime import *
 from main.models import PoemaModel
 from main.models import UsuarioModel
 from main.models import CalificacionModel
-
+from main.auth.decoradores import admin_required
+from flask_jwt_extended import jwt_required, get_jwt_identity, get_jwt
 class Poema(Resource):
+    @jwt_required()
     def get(self, id):
         poema = db.session.query(PoemaModel).get_or_404(id)
         return poema.to_json()
-
+    @jwt_required()
     def delete(self, id):
+        claims = get_jwt()
+        id_usuario = get_jwt_identity()
         poema = db.session.query(PoemaModel).get_or_404(id)
-        db.session.delete(poema)
-        db.session.commit()
-        return '', 204
+        if "rol" in claims:
+            if claims['rol'] == "admin" or id_usuario == int(poema.usuarioid):
+                db.session.delete(poema)
+                db.session.commit()
+                return '', 204
+            else:
+                return "Este usuario no puede realizar esa acción"
 
     
 
@@ -77,9 +85,20 @@ class Poemas(Resource):
     
 
    
-
+    @jwt_required()
     def post(self):
+        id_usuario = get_jwt_identity()
         poema = PoemaModel.from_json(request.get_json())
-        db.session.add(poema)
-        db.session.commit()
-        return poema.to_json(), 201
+        usuario = db.session.query(UsuarioModel).get_or_404(id_usuario)
+        claims = get_jwt()
+        if "rol" in claims:
+            if claims["rol"] == "poeta":
+                if len(usuario.poemas) == 0 or len(usuario.calificaciones) >= 2:
+                    poema.usuario_id = id_usuario
+                    db.session.add(poema)
+                    db.session.commit()
+                    return poema.to_json(), 201
+                else:
+                    return "No hay suficientes calificaciones por parte de este usuario"
+            else:
+                return "Este usuario no puede realizar esta acción."
